Add connection stats debug mode to transmitter.
When using this mode the receiver should be in passive mode (undef all ESC
modes) but still powered on in the board. A different remote and receiver should
be used to ride, with the feather remote in pocket or offhand. The transmitter
will show a histogram of command delays, and alert the user when connection
issues are detected.